+++ zihanki-status.rb
